fix: 버그 수정, logging 수정, DTO 수정

+++ gadget/src/anchor/mod.rs
@@ -35,7 +35,7 @@ pub trait AnchorScheme {
     fn generate_witness(
         pk: &Self::PublicKey,
         secrets: &Self::Secret,
-        selector: &[usize],
+        selector: &[u8],
         matrix: &Self::Matrix,
     ) -> Result<Self::Witness, AnchorError>;
